Move examples to another file
